+++ ScoreManagement/Model/ExcelScore/ExcelScoreModel.cs
@@ -9,9 +9,9 @@
         public string? Section { get; set; }
         public string? ScoreType { get; set; }
         public int? StudentCount { get; set; }
-        public int? AverageScore { get; set; }
-        public int? MinScore { get; set; }
-        public int? MaxScore { get; set; }
+        public decimal? AverageScore { get; set; }
+        public decimal? MinScore { get; set; }
+        public decimal? MaxScore { get; set; }
         public double? StandardDeviation { get; set; }
         public int? Sum0_39 { get; set; }
         public int? Sum40_49 { get; set; }
